# api/openapi.yaml
openapi: 3.1.0
info:
  title: Vermouth gateway
  version: 0.1.0
  description: |
    The single description of the gateway surface (STK-10). The gateway's Go
    request and response types are generated from this document with
    oapi-codegen, and the browser's types with openapi-typescript. A hand
    written client type for a gateway endpoint is not allowed.

    Only the gateway is described here. Services are reached by the gateway
    alone (INV-1), so their endpoints are not a public surface.

    The four auth endpoints are unauthenticated because they are what produces
    a token. Everything else under /api/ needs one. The teaching endpoints are
    the first real product thread through the gateway.
servers:
  - url: http://localhost:8080
    description: local development
paths:
...
  /api/classes:
    post:
      operationId: createClass
      summary: Create a class with its first session
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClassRequest"
      responses:
        "200":
          description: The original class and first session for a safe retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClassResponse"
        "201":
          description: The class and first session were created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateClassResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/classes/{class_id}/schedule:
    put:
      operationId: putClassSchedule
      summary: Create the first weekly schedule for an existing class
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutScheduleRequest"
      responses:
        "200":
          description: A replacement rule was created or the original response was replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutScheduleResponse"
        "201":
          description: The first retained schedule rule and concrete sessions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutScheduleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/classes/{class_id}/rates:
    get:
      operationId: getClassRates
      summary: Read authoritative current rate and projected dated history
      parameters:
        - $ref: "#/components/parameters/ClassId"
      responses:
        "200":
          description: Current rate, allowed correction range, and complete dated history
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassRates"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/classes/{class_id}/rates/{effective_date}:
    put:
      operationId: putClassRate
      summary: Create or correct one dated class rate
      parameters:
        - $ref: "#/components/parameters/ClassId"
        - name: effective_date
          in: path
          required: true
          schema:
            type: string
            format: date
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PutClassRateRequest"
      responses:
        "200":
          description: The committed dated rate or its original retry response
          headers:
            Cache-Control:
              required: true
              schema:
                type: string
                const: no-store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PutClassRateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/classes/{class_id}/schedule/end:
    post:
      operationId: endClassSchedule
      summary: End the latest retained weekly schedule
      parameters:
        - name: class_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EndScheduleRequest"
      responses:
        "200":
          description: The schedule was ended or the original response was replayed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EndScheduleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/sessions/{session_id}/move:
    post:
      operationId: moveSession
      summary: Move one active session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoveSessionRequest"
      responses:
        "200":
          description: The moved canonical session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanonicalSession"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/sessions/{session_id}/cancel:
    post:
      operationId: cancelSession
      summary: Cancel one active session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionVersionRequest"
      responses:
        "200":
          description: The cancelled canonical session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanonicalSession"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/sessions/{session_id}/restore:
    post:
      operationId: restoreSession
      summary: Restore one tutor cancelled session
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionVersionRequest"
      responses:
        "200":
          description: The restored canonical session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CanonicalSession"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/schedule:
    get:
      operationId: getSchedule
      summary: Read one bounded tutor calendar window
      parameters:
        - name: from
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: through
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: class_id
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              format: uuid
        - name: include_replaced
          in: query
          required: false
          schema:
            type: boolean
            default: false
        - name: history_limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 50
        - name: history_cursor
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: The owned classes, rules, and concrete sessions in the display window
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Schedule"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
  /api/students:
    get:
      operationId: listStudents
      summary: Search active students
      parameters:
        - $ref: "#/components/parameters/StudentQuery"
        - $ref: "#/components/parameters/StudentCursor"
      responses:
        "200":
          description: One stable cursor page of active students
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
    post:
      operationId: createStudent
      summary: Register a student
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
      responses:
        "200":
          description: The original student for a safe retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"
        "201":
          description: The student was registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/students/{student_id}:
    get:
      operationId: getStudent
      summary: Read one active student and class history
      parameters:
        - $ref: "#/components/parameters/StudentId"
      responses:
        "200":
          description: The active student and retained memberships
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateStudent
      summary: Edit an active student using the last read timestamp
      parameters:
        - $ref: "#/components/parameters/StudentId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentRequest"
      responses:
        "200":
          description: The canonical updated student
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      operationId: archiveStudent
      summary: Archive an active student with no current class membership
      parameters:
        - $ref: "#/components/parameters/StudentId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: The student was archived or the exact command was replayed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/classes/{class_id}/roster:
    get:
      operationId: getClassRoster
      summary: Read an owned class roster on a local date
      parameters:
        - $ref: "#/components/parameters/ClassId"
        - $ref: "#/components/parameters/RosterDate"
      responses:
        "200":
          description: The owned class and students covered by the resolved date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassRoster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: changeClassRoster
      summary: Apply one atomic dated roster delta
      parameters:
        - $ref: "#/components/parameters/ClassId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangeRosterRequest"
      responses:
        "200":
          description: The canonical roster on the change date
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassRoster"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/sessions/{session_id}/attendance:
    get:
      operationId: getSessionAttendance
      summary: Read one whole roster attendance sheet
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: The coherent session, roster, saved states, and revision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSheet"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: saveSessionAttendance
      summary: Save one complete whole roster attendance pass
      parameters:
        - $ref: "#/components/parameters/SessionId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAttendanceRequest"
      responses:
        "200":
          description: The complete canonical saved pass
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSave"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /api/home:
    get:
      operationId: getHome
      summary: The tutor and today's teaching work
      parameters:
        - $ref: "#/components/parameters/HomeCursor"
      responses:
        "200":
          description: Required identity and teaching facts plus optional billing progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Home"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "502":
          $ref: "#/components/responses/BadGateway"
...
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: One browser generated UUID, retained until this create step succeeds.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    ClassId:
      name: class_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    StudentCursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor bound to the trimmed student search.
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: An Ed25519 signed access token, verified locally by each service (INV-14).
  responses:
    BadRequest:
      description: The request was not acceptable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthenticated:
      description: No valid token was supplied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: No owned resource matched
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The requested state conflicts with committed state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    CreateClassRequest:
      type: object
      description: Exactly one of first_session or schedule is required.
      required: [name, rate_amount]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        color:
          oneOf:
            - $ref: "#/components/schemas/ClassColor"
            - type: "null"
        rate_amount:
          type: integer
          format: int64
          minimum: 0
          maximum: 1000000000
        first_session:
          $ref: "#/components/schemas/FirstSessionInput"
        schedule:
          $ref: "#/components/schemas/WeeklyScheduleInput"
...
    CreateClassResponse:
      type: object
      required: [class, first_session, rule, candidate_count, created_count, adopted_count]
      properties:
        class:
          $ref: "#/components/schemas/Class"
        first_session:
          oneOf:
            - $ref: "#/components/schemas/TeachingSession"
            - type: "null"
        rule:
          oneOf:
            - $ref: "#/components/schemas/ScheduleRuleSummary"
            - type: "null"
        candidate_count:
          type: integer
          minimum: 0
        created_count:
          type: integer
          minimum: 0
        adopted_count:
          type: integer
          minimum: 0
    PutScheduleRequest:
      type: object
      required: [expected_revision, effective_from, valid_through, slots]
      properties:
        expected_revision:
          type: integer
          format: int64
          minimum: 0
        effective_from:
          type: string
          format: date
        valid_through:
          type: string
          format: date
        slots:
          type: array
          minItems: 1
          maxItems: 7
          items:
            $ref: "#/components/schemas/WeeklyScheduleSlotInput"
    PutScheduleResponse:
      type: object
      required:
        [class, rule, first_session, candidate_count, created_count, adopted_count, superseded_count, preserved_count]
      properties:
        class:
          $ref: "#/components/schemas/Class"
        rule:
          $ref: "#/components/schemas/ScheduleRuleSummary"
        first_session:
          oneOf:
            - $ref: "#/components/schemas/TeachingSession"
            - type: "null"
        candidate_count:
          type: integer
          minimum: 0
        created_count:
          type: integer
          minimum: 0
        adopted_count:
          type: integer
          minimum: 0
        superseded_count:
          type: integer
          minimum: 0
        preserved_count:
          type: integer
          minimum: 0
...
security:
  - bearerAuth: []